Add type statement, new() function
type n {...} creates a structure definition and names it n, but also
stores metadata in the struct that identifies it as a type.

new(n) creates a deep copy of n and return it as a new struct, but it
has a __type linkage that points to the original struct, not its name.

This allows a.b references to use the __type linkage to find type values