bug: io.expand() extension should act as filter
Using the optional second parameter in io.expand() will try to
add it as the extension to a filename to see if it can be read.
This fixes the case where, if specified, it also acts as filter
when the path is a directory; only files ending in that string
are reported.