feat: make symbol name normalization optional
Previously, most symbol names were coerced to lower-case. This change
means the default is that symbol names are unchanged (and must match
exactly to be found in an expression reference, etc.).

Added a LowercaseIdentifiers flag to the compiler structure that can
be set after a compiler.New() operation.

Added a persistence.GetBool() function which reads a key, and parses
the value to determine if it reflects "true" or "false" values, and
returns the boolean result.