Add flag to AutoImport to control required vs optional packages
The compiler.AutoImport() method should be called any time a new
compiler object is being used to run user code. This used to be
optional. However, healthy ego sessions require at least the
util package, and profile helps them determine their state.

With this change, you pass a flag to AutoImport indicating if you
want all packages (true) or just the required packages (false).
This change also adds a global array which is the list of required
packages, initialized to ["util", "profile"] but can be extended
in the future if needed.