Add ability to invoke function without "call" verb
This makes Ego more "Go-like", so you don't have to do this:

    call profile.set("test", "true")

but can instead do this:

    profile.set("test", "true")

The compiler now detects when the first part of a statement is a
function call, and invokes it.

This also adds a new package, "fmt", with printf(), println(), and
sprintf() functions patterned after the Go equivalents.